requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
Flask==2.1.1
idna==3.3
imbalanced-learn==0.9.0
imblearn==0.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
MarkupSafe==2.1.1
numpy==1.22.3
python-dotenv==0.20.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
spotipy==2.19.0
threadpoolctl==3.1.0
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
gunicorn==20.1.0
